cloudinary,like feature, profile feature implemented

+++ src/models/likeModel.js
@@ -13,15 +13,24 @@ const likeModel = {
     return rows[0];
   },
 
-  async removeLike(userId, postId) {
-    const query = `
-      DELETE FROM likes
-      WHERE user_id = $1 AND post_id = $2
-      RETURNING *;
-    `;
-    const { rows } = await db.query(query, [userId, postId]);
-    return rows[0];
+  async countLikes(postId){
+    const result = await db.query(
+      `SELECT COUNT(*) FROM likes WHERE post_id = $1`,
+      [postId]
+    );
+    return parseInt(result.rows[0].count);
   },
+  async hasLiked(userId,postId){
+    const result=await db.query(`select * from likes where user_id=$1 AND post_id=$2`,[userId,postId]);
+    return  result.rows.length>0;
+  },
+  async removeLike(userId, postId) {
+  const result = await db.query(
+    `DELETE FROM likes WHERE user_id = $1 AND post_id = $2 RETURNING *`,
+    [userId, postId]
+  );
+  return result.rowCount > 0; // true if deleted
+},
 
   async getLikesByPost(postId) {
     const query = `
@@ -39,7 +48,7 @@ const likeModel = {
     `;
     const { rowCount } = await db.query(query, [userId, postId]);
     return rowCount > 0;
-  }
+  } 
 };
 
 module.exports = likeModel;
